Chapter 9 Removing ingredients when removing a bookmark

+++ 09-room-db/projects/final/app/src/main/java/com/kodeco/recipefinder/data/Repository.kt
@@ -78,7 +78,7 @@ class Repository(recipeDatabase: RecipeDatabase) {
     recipeDao.deleteRecipe(recipe)
   }
 
-  suspend fun deleteRecipeById(recipeId: Long) {
+  suspend fun deleteRecipeById(recipeId: Int) {
     recipeDao.deleteRecipeById(recipeId)
   }
